backend/sicargabox/requirements.txt
# Core Framework
Django==5.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
channels==4.3.1

# Database & Search
psycopg==3.2.10
psycopg-binary==3.2.10
# django-pgvector==0.2.5 # Removed as we use Elasticsearch now

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
django-guardian==3.2.0

# API & Data Processing
pandas==2.3.3
numpy==2.3.3
openpyxl==3.1.2
python-dotenv==1.0.1

# API Documentation & Filtering
drf-spectacular==0.28.0
django-filter==25.2

# AI & Machine Learning
openai==1.12.0
deepseek==1.0.0
anthropic==0.49.0
tiktoken==0.9.0

# Development & Testing
pytest==8.0.2
pytest-django==4.8.0
pytest-cov==4.1.0
black==24.2.0
flake8==7.0.0
isort==5.13.2
mypy==1.18.2
django-stubs==5.2.7
types-requests==2.32.4.20250913
types-urllib3==1.26.25.14

# Documentation
drf-yasg==1.21.7
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
requests==2.32.5
urllib3==2.5.0
Pillow==10.5.1
tqdm==4.67.1

# Forms / Templates
django-widget-tweaks==1.5.0
django-select2==8.1.2
django-tailwind==3.8.0
django-htmx==1.26.0

# Development Tools
django-debug-toolbar==6.0.0

# Search
elasticsearch==8.17.0
elasticsearch-dsl==8.17.1
django-elasticsearch-dsl==8.0
